Windows errno_to_str: explicit casts to clear warnings

+++ src/internal_modules/roc_core/target_windows/roc_core/errno_to_str.cpp
@@ -13,7 +13,7 @@ namespace roc {
 namespace core {
 
 errno_to_str::errno_to_str() {
-    format_(GetLastError());
+    format_((int)GetLastError());
 }
 
 errno_to_str::errno_to_str(int err) {
@@ -25,8 +25,8 @@ void errno_to_str::format_(int err) {
 
     size_t size =
         FormatMessageW(FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
-                       nullptr, err, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), wbuf,
-                       sizeof(wbuf) / sizeof(wchar_t) - 1, nullptr);
+                       nullptr, (DWORD)err, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
+                       wbuf, sizeof(wbuf) / sizeof(wchar_t) - 1, nullptr);
 
     if (size == 0) {
         strcpy(buffer_, "<unknown error>");
